add 'new shared feed' button anf refactor component
