+++ src/app/components/drawer/FeedPage.module.css
@@ -72,3 +72,7 @@
   gap: 4px;
   margin: 10px;
 }
+
+.createFeedButton {
+  margin: 0 32px;
+}
